refactor(llm): extract litellm bootstrap + structured-content shim to _litellm_shims
Moves the LiteLLM bootstrap layer out of provider.py into a new
amx/llm/_litellm_shims.py:

- _configure_ssl_environment + AMX_INSECURE_SSL env handling
- _litellm (lazy module getter + cost-map env hint + logger silencer)
- _install_structured_content_shim (two-layer monkeypatch on
  litellm.utils.convert_to_model_response_object AND Message.__init__)
- _normalize_response_dict_in_place + _flatten_structured_content
  (the actual list-shaped content -> str flatteners)
- the _litellm_module global lives here too, since the shim install
  uses it as the once-per-session guard

provider.py re-exports each public name so call sites that patch the
shims still work. Two regression tests (test_litellm_silenced.py and
test_regressions.LLMProviderTests) had their patch path updated to
amx.llm._litellm_shims._litellm_module since the global moved.

provider.py: 1244 -> 1017 LOC (-227, -18%). Cumulative across PRs
#470 + #471 + #472 + this: 1733 -> 1017 LOC (-41%).
